+++ src/static-pages/_LoadingPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
+import ApplicationLoadingOverlay from 'terra-application/lib/application-loading-overlay';
 
 // Wait half a second before showing the loading indicator.
 const LoadingPage = () => {
@@ -15,7 +15,7 @@ const LoadingPage = () => {
     return () => { isActive = false; };
   }, []);
 
-  return <LoadingOverlay message="" isOpen={state.isOpen} isAnimated isRelativeToContainer zIndex="6000" />;
+  return <ApplicationLoadingOverlay isOpen={state.isOpen} />;
 };
 
 export default LoadingPage;